# APP/contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/admin/artifacts/{artifactId}": {
      "get": {
        "operationId": "AdminArtifactsController_getMetadata",
        "parameters": [
          {
            "name": "artifactId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminArtifacts"
        ]
      },
      "delete": {
        "operationId": "AdminArtifactsController_delete",
        "parameters": [
          {
            "name": "artifactId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "AdminArtifacts"
        ]
      }
    },
    "/api/admin/artifacts/{artifactId}/download-url": {
      "get": {
        "operationId": "AdminArtifactsController_getDownloadUrl",
        "parameters": [
          {
            "name": "artifactId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminArtifacts"
        ]
      }
    },
    "/api/admin/artifacts/{artifactId}/content": {
      "get": {
        "operationId": "AdminArtifactsController_getContent",
        "parameters": [
          {
            "name": "artifactId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminArtifacts"
        ]
      }
    },
    "/api/groups/{groupId}/artifacts": {
      "get": {
        "operationId": "ArtifactsController_listArtifacts",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Artifacts"
        ]
      }
    },
    "/api/groups/{groupId}/artifacts/confirm": {
      "post": {
        "operationId": "ArtifactsController_confirmArtifacts",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Artifacts"
        ]
      },
      "delete": {
        "operationId": "ArtifactsController_unlockArtifacts",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Artifacts"
        ]
      }
    },
    "/api/groups/{groupId}/artifacts/from-files/{fileId}": {
      "post": {
        "operationId": "ArtifactsController_addFileToArtifacts",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fileId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinalizedArtifactValueResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Artifacts"
        ]
      }
    },
    "/api/groups/{groupId}/artifacts/{artifactId}": {
      "get": {
        "operationId": "ArtifactsController_getArtifactMetadata",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "artifactId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Artifacts"
        ]
      },
      "delete": {
        "operationId": "ArtifactsController_deleteArtifact",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "artifactId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Artifacts"
        ]
      }
    },
    "/api/groups/{groupId}/artifacts/{artifactId}/download-url": {
      "get": {
        "operationId": "ArtifactsController_getArtifactDownloadUrl",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "artifactId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Artifacts"
        ]
      }
    },
    "/api/groups/{groupId}/artifacts/{artifactId}/content": {
      "get": {
        "operationId": "ArtifactsController_getArtifactContent",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "artifactId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Artifacts"
        ]
      }
    },
...
    "/api/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "auth-legacy-reminder"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "auth-legacy-reminder"
        ]
      }
    },
    "/api/auth/request-code": {
      "post": {
        "operationId": "AuthController_requestCode",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestEmailCodeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "auth-legacy-reminder"
        ]
      }
    },
    "/api/auth/verify-code": {
      "post": {
        "operationId": "AuthController_verifyCode",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailCodeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "auth-legacy-reminder"
        ]
      }
    },
...
    "/api/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshSessionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrowserSessionResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth-legacy-reminder"
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshSessionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth-legacy-reminder"
        ]
      }
    },
...
    "/api/auth/token/register": {
      "post": {
        "operationId": "TokenAuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenSessionResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth-token"
        ]
      }
    },
    "/api/auth/token/login": {
      "post": {
        "operationId": "TokenAuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenSessionResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth-token"
        ]
      }
    },
    "/api/auth/token/request-code": {
      "post": {
        "operationId": "TokenAuthController_requestCode",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestEmailCodeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestCodeResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth-token"
        ]
      }
    },
    "/api/auth/token/verify-code": {
      "post": {
        "operationId": "TokenAuthController_verifyCode",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailCodeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenSessionResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth-token"
        ]
      }
    },
    "/api/auth/token/refresh": {
      "post": {
        "operationId": "TokenAuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshSessionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenSessionResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth-token"
        ]
      }
    },
    "/api/auth/token/logout": {
      "post": {
        "operationId": "TokenAuthController_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshSessionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth-token"
        ]
      }
    },
...
  },
  "info": {
    "title": "SekerChat HTTP API",
    "description": "Published contract for browser, CLI, reminder, and mobile clients.",
    "version": "0.8.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "seker_access"
      }
    },
    "schemas": {
      "FinalizedArtifactValueResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "groupId": {
            "type": "string"
          },
          "uploaderId": {
            "type": "string"
          },
          "originalName": {
            "type": "string"
          },
          "storedName": {
            "type": "string"
          },
          "relativePath": {
            "type": "string"
          },
          "mimeType": {
            "type": "string"
          },
          "size": {
            "type": "number"
          },
          "sourceFileId": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string"
          },
          "contentUrl": {
            "type": "string"
          },
          "metadataUrl": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "groupId",
          "uploaderId",
          "originalName",
          "storedName",
          "relativePath",
          "mimeType",
          "size",
          "sourceFileId",
          "createdAt",
          "contentUrl",
          "metadataUrl"
        ]
      },
...
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "pattern": "(?=.*[A-Z])"
          },
          "displayName": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RequestEmailCodeDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          }
        },
        "required": [
          "email"
        ]
      },
      "VerifyEmailCodeDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "code": {
            "type": "string",
            "minLength": 6,
            "maxLength": 6
          }
        },
        "required": [
          "email",
          "code"
        ]
      },
...
      "AuthUserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "type": "string"
          },
          "mustChangePassword": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "email",
          "displayName",
          "role"
        ]
      },
...
      "LogoutResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ]
      },
      "RefreshSessionDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "minLength": 10
          }
        }
      },
      "BrowserSessionResponseDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/AuthUserResponseDto"
          }
        },
        "required": [
          "user"
        ]
      },
...
      "RequestCodeResponseDto": {
        "type": "object",
        "properties": {
          "deliveryHint": {
            "type": "string"
          }
        },
        "required": [
          "deliveryHint"
        ]
      },
      "TokenSessionResponseDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/AuthUserResponseDto"
          },
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "user",
          "accessToken",
          "refreshToken"
        ]
      },
...
    }
  }
}